Change Config trait to MdsHandler.
