+++ common/src/lib.rs
@@ -9,7 +9,7 @@ pub mod consts;
 pub mod error;
 pub mod event;
 pub mod timer;
-//pub mod channel;
+pub mod method;
 pub mod uds_server;
 pub mod uds_client;
 pub mod acl;
