+++ routerd/src/core/mds.rs
@@ -2,11 +2,9 @@
 // ReZe.Rs - Router Daemon
 //   Copyright (C) 2018-2020 Toshiaki Takada
 //
-// Core - Config.
+// Core - Manageement Data Store.
 //
 
-use std::fmt;
-use std::str::FromStr;
 use std::rc::Rc;
 
 use log::debug;
@@ -14,65 +12,22 @@ use regex::Regex;
 
 use common::error::*;
 
-/// Method: equivalent to HTTP Method.
-#[derive(Copy, Clone)]
-pub enum Method {
-    Get,
-    Post,
-    Put,
-    Delete,
-    Patch,
-}
-
-/// FromStr.
-impl FromStr for Method {
-    type Err = CoreError;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        let method = s.to_lowercase();
-
-        match method.as_ref() {
-            "get" => Ok(Method::Get),
-            "post" => Ok(Method::Post),
-            "put" => Ok(Method::Put),
-            "delete" => Ok(Method::Delete),
-            "patch" => Ok(Method::Patch),
-            _ => Err(CoreError::ParseMethod),
-        }
-    }
-}
-
-/// Display.
-impl fmt::Display for Method {
-
-    /// Format method.
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        let s = match self {
-            Method::Get => "GET",
-            Method::Post => "POST",
-            Method::Put => "PUT",
-            Method::Delete => "DELETE",
-            Method::Patch => "PATCH",
-        };
-
-        write!(f, "{}", s)
-    }
-}
-
-/// Config trait.
-pub trait Config {
+/// Management Data Store trait.
+///  Store config or device state in hierarchy.
+///  Dispatch REST like request to certain data objects.
+pub trait MdsHandler {
 
     /// Return unique identifier, this is used to register to parent as a key.
     fn id(&self) -> &str;
 
-    /// Register child Config trait object to this Config.
-    fn register_child(&mut self, _child: Rc<dyn Config>) {
-        debug!("This object does not have child Config object");
+    /// Register child MdsHandler trait object to this Config.
+    fn register_child(&mut self, _child: Rc<dyn MdsHandler>) {
+        debug!("This object does not have child MdsHandler object");
         ()
     }
 
-    /// Lookup child Config with given path.
-    fn lookup_child(&self, _path: &str) -> Option<Rc<dyn Config>> {
+    /// Lookup child MdsHandler with given path.
+    fn lookup_child(&self, _path: &str) -> Option<Rc<dyn MdsHandler>> {
         debug!("Not implemented");
         None
     }
