+++ routerd/src/core/nexus.rs
@@ -30,6 +30,7 @@ use log::error;
 use common::event::*;
 use common::timer::*;
 use common::error::*;
+use common::method::Method;
 use common::uds_server::*;
 
 use super::signal;
@@ -40,7 +41,7 @@ use super::message::nexus::NexusToProto;
 use super::message::zebra::ProtoToZebra;
 use super::message::zebra::ZebraToProto;
 use super::master::ProtocolMaster;
-use super::config::*;
+use super::mds::*;
 use super::config_master::*;
 
 use crate::zebra::master::ZebraMaster;
